+++ src/Plugin/src/Export/KmlExport/src/KmlExportPlugin.h
@@ -59,7 +59,9 @@ protected:
     virtual QString getFileExtension() const noexcept override;
     virtual QString getFileFilter() const noexcept override;
     virtual std::unique_ptr<QWidget> createOptionWidget() const noexcept override;
+    virtual bool hasMultiAircraftSupport() const noexcept override;
     virtual bool exportFlight(const Flight &flight, QIODevice &io) noexcept override;
+    virtual bool exportAircraft(const Flight &flight, const Aircraft &aircraft, QIODevice &io) noexcept override;
 
 protected slots:
     virtual void onRestoreDefaultSettings() noexcept override;