+++ src/Plugin/src/ExportPluginBase.cpp
@@ -54,6 +54,8 @@ class ExportPluginBasePrivate
 public:
     ExportPluginBasePrivate()
     {}
+
+    std::vector<QString> exportedFilePaths;
 };
 
 // PUBLIC
@@ -75,7 +77,6 @@ ExportPluginBase::~ExportPluginBase() noexcept
 
 bool ExportPluginBase::exportFlight(const Flight &flight) noexcept
 {
-    bool ok;
     std::unique_ptr<QWidget> optionWidget = createOptionWidget();
     ExportPluginBaseSettings &baseSettings = getSettings();
     std::unique_ptr<BasicExportDialog> exportDialog = std::make_unique<BasicExportDialog>(flight, getFileExtension(), getFileFilter(), baseSettings, getParentWidget());
@@ -83,6 +84,7 @@ bool ExportPluginBase::exportFlight(const Flight &flight) noexcept
             this, &ExportPluginBase::onRestoreDefaultSettings);
     // Transfer ownership to exportDialog
     exportDialog->setOptionWidget(optionWidget.release());
+    bool ok {true};
     const int choice = exportDialog->exec();
     if (choice == QDialog::Accepted) {
         // Remember export path
@@ -93,7 +95,8 @@ bool ExportPluginBase::exportFlight(const Flight &flight) noexcept
             const QString exportDirectoryPath = fileInfo.absolutePath();
             Settings::getInstance().setExportPath(exportDirectoryPath);
 
-            if (baseSettings.isFileDialogSelectedFile() || !fileInfo.exists()) {
+            const ExportPluginBaseSettings::FormationExport formationExport = getSettings().getFormationExport();
+            if (formationExport == ExportPluginBaseSettings::FormationExport::AllAircraftSeparateFiles || baseSettings.isFileDialogSelectedFile() || !fileInfo.exists()) {
                 ok = exportFlight(flight, filePath);
             } else {
                 std::unique_ptr<QMessageBox> messageBox = std::make_unique<QMessageBox>(getParentWidget());
@@ -115,8 +118,6 @@ bool ExportPluginBase::exportFlight(const Flight &flight) noexcept
         } else {
             ok = true;
         }
-    } else {
-        ok = true;
     }
 
     return ok;
@@ -150,32 +151,106 @@ void ExportPluginBase::resamplePositionDataForExport(const Aircraft &aircraft, s
 
 bool ExportPluginBase::exportFlight(const Flight &flight, const QString &filePath) noexcept
 {
+    d->exportedFilePaths.clear();
     QFile file(filePath);
-    bool ok = file.open(QIODevice::WriteOnly);
-    if (ok) {
+    bool ok {true};
 #ifdef DEBUG
-        QElapsedTimer timer;
-        timer.start();
+    QElapsedTimer timer;
+    timer.start();
 #endif
-        QGuiApplication::setOverrideCursor(Qt::WaitCursor);
-        QGuiApplication::processEvents();
-        ok = exportFlight(flight, file);
-        QGuiApplication::restoreOverrideCursor();
-#ifdef DEBUG
-        qDebug("%s export %s in %lld ms", qPrintable(QFileInfo(filePath).fileName()), (ok ? qPrintable("SUCCESS") : qPrintable("FAIL")), timer.elapsed());
-#endif
+    QGuiApplication::setOverrideCursor(Qt::WaitCursor);
+    QGuiApplication::processEvents();
+    const ExportPluginBaseSettings &settings = getSettings();
+    switch (settings.getFormationExport()) {
+    case ExportPluginBaseSettings::FormationExport::UserAircraftOnly:
+        ok = file.open(QIODevice::WriteOnly);
+        if (ok) {
+            ok = exportAircraft(flight, flight.getUserAircraftConst(), file);
+            d->exportedFilePaths.push_back(filePath);
+        }
+        file.close();
+        break;
+    case ExportPluginBaseSettings::FormationExport::AllAircraftOneFile:
+        if (hasMultiAircraftSupport()) {
+            ok = file.open(QIODevice::WriteOnly);
+            if (ok) {
+                ok = exportFlight(flight, file);
+                d->exportedFilePaths.push_back(filePath);
+            }
+            file.close();
+        } else {
+            ok = exportAllAircraft(flight, filePath);
+        }
+        break;
+    case ExportPluginBaseSettings::FormationExport::AllAircraftSeparateFiles:
+        ok = exportAllAircraft(flight, filePath);
+        break;
+    default:
+        ok = false;
+        break;
     }
-
-    file.close();
+    QGuiApplication::restoreOverrideCursor();
+#ifdef DEBUG
+    qDebug("%s export %s in %lld ms", qPrintable(QFileInfo(filePath).fileName()), (ok ? qPrintable("SUCCESS") : qPrintable("FAIL")), timer.elapsed());
+#endif
 
     if (ok) {
-        if (getSettings().isOpenExportedFileEnabled()) {
-            const QString fileUrl = QString("file:///") + filePath;
-            QDesktopServices::openUrl(QUrl(fileUrl));
+        if (settings.isOpenExportedFilesEnabled()) {
+            for (const QString &exportedFilePath : d->exportedFilePaths) {
+                const QString fileUrl = QString("file:///") + exportedFilePath;
+                QDesktopServices::openUrl(QUrl(fileUrl));
+            }
         }
     } else {
         QMessageBox::critical(getParentWidget(), tr("Export error"), tr("An error occured during export into file %1.").arg(filePath));
     }
+
+    return ok;
+}
+
+bool ExportPluginBase::exportAllAircraft(const Flight &flight, const QString &filePath) noexcept
+{
+    bool ok {true};
+    bool replaceAll {false};
+    int i = 1;
+    for (const auto &aircraft : flight) {
+        // @todo Generate enumerated file names, check existence for each before overwriting!
+        const QString sequencedFilePath = File::getSequenceFilePath(filePath, i);
+        const QFileInfo fileInfo {sequencedFilePath};
+        if (fileInfo.exists() && !replaceAll) {
+            QGuiApplication::restoreOverrideCursor();
+            std::unique_ptr<QMessageBox> messageBox = std::make_unique<QMessageBox>(getParentWidget());
+            messageBox->setIcon(QMessageBox::Question);
+            QPushButton *replaceButton = messageBox->addButton(tr("&Replace"), QMessageBox::AcceptRole);
+            QPushButton *replaceAllButton = messageBox->addButton(tr("Replace &all"), QMessageBox::YesRole);
+            messageBox->setText(tr("A file named \"%1\" already exists. Do you want to replace it?").arg(fileInfo.fileName()));
+            messageBox->setInformativeText(tr("The file already exists in \"%1\".  Replacing it will overwrite its contents.").arg(fileInfo.dir().dirName()));
+            messageBox->setStandardButtons(QMessageBox::Cancel);
+            messageBox->setDefaultButton(replaceButton);
+
+            messageBox->exec();
+            const QAbstractButton *clickedButton = messageBox->clickedButton();
+            if (clickedButton == replaceAllButton) {
+                replaceAll = true;
+            } else if (clickedButton != replaceButton) {
+                break;
+            }
+            QGuiApplication::setOverrideCursor(Qt::WaitCursor);
+            QGuiApplication::processEvents();
+        }
+
+        QFile file(sequencedFilePath);
+        ok = file.open(QIODevice::WriteOnly);
+        if (ok) {
+            ok = exportAircraft(flight, *aircraft, file);
+            d->exportedFilePaths.push_back(sequencedFilePath);
+        }
+        file.close();
+        ++i;
+        if (!ok) {
+            break;
+        }
+    } // All aircraft
 
     return ok;
 }